capture all routes to show 404 and add a character detail not found
