+++ src/App.css
@@ -1,0 +1,3 @@
+::-webkit-scrollbar {
+  display: none;
+}
